requirements.txt
fastapi
fastapi-cache2
numpy
psycopg2-binary
pydantic
pydantic_core
SQLAlchemy
uvicorn
requests
dotenv
cachetools
python-dotenv
pytz
redis